allow supply credentials with env reference
Allow the user to specify environment reference for secret cnfiguretions.
If use specifies value like `$ENV_NAME` or `${ENV_NAME}`, we will
try to replace the value with the value of the environment variable.

For AWS credentials, also try to read value from common AWS credential
environment variables if available.